非同期通信の実装、時間表示を日本対応にするgemを導入
