+++ app/javascript/answer.js
@@ -1,25 +1,15 @@
-import axios from 'axios';
-window.addEventListener("load", () => {
-  const submit = document.getElementById("submit");
-  submit.addEventListener("click", (e) => {
-    const formData = new FormData(document.getElementById("form"));
-    e.preventDefault();
-    axios.post('/answers', formData)
-      .then(function (res) {
-        const item = res.data.post;
-        const list = document.getElementById("list");
-        const formText = document.getElementById("answer");
+window.addEventListener('load', () => {
+    const incBtn = document.getElementById("incbtn")
+    incBtn.addEventListener('click', (e) => {
+        e.preventDefault();
+        const addForm = document.getElementById("addinc")
         const HTML = `
-                  <div class="post" data-id=${item.id}>
-                    <div class="post-date">
-                      投稿日時：${item.created_at}
-                    </div>
-                    <div class="post-content">
-                    ${item.answer}
-                    </div>
-                  </div>`;
-        list.insertAdjacentHTML("afterend", HTML);
-        formText.value = "";
-      });
-  });
+        <div class="right">
+            <form action="" , id="inner">
+                <input type="text", id="pra", value="">
+                <input type="button", id="btn", value="送信">
+            </form>
+        </div >`;
+        addForm.insertAdjacentHTML("afterend", HTML);
+    });
 })
