+++ app/views/shared/_header.html.erb
@@ -27,7 +27,7 @@
   <%# </ul> %>
   <%= link_to 'ホームへ戻る', root_path, class: "root-back" %>
   <%= link_to '満足度ページ', satisfactions_path, method: :get, class: "root-back" %>
-  <%= link_to '分析一覧にいく', answer_path(current_user), method: :get, class: "login" %>
+  
 
   <ul class='lists-right'>
     <%# deviseを導入できたら、ログインの有無で表示が変わるように分岐しましょう%>
